requirements.txt
cryptography==44.0.0
fastapi==0.115.6
uvicorn
pydantic==2.10.4
PyJWT==2.10.1
pymongo~=3.12.0
python-dotenv==1.0.1
certifi~=2024.12.14